[#2750] Move get_value to ProcedurePresentation
