+++ app/models/procedure_presentation.rb
@@ -52,6 +52,23 @@ class ProcedurePresentation < ApplicationRecord
     end
   end
 
+  def get_value(dossier, table, column)
+    dossier_field_service.assert_valid_column(dossier.procedure, table, column)
+
+    case table
+    when 'self'
+      dossier.send(column)
+    when 'user'
+      dossier.user.send(column)
+    when 'etablissement'
+      dossier.etablissement&.send(column)
+    when 'type_de_champ'
+      dossier.champs.find { |c| c.type_de_champ_id == column.to_i }.value
+    when 'type_de_champ_private'
+      dossier.champs_private.find { |c| c.type_de_champ_id == column.to_i }.value
+    end
+  end
+
   def sorted_ids(dossiers, gestionnaire)
     table = sort['table']
     column = sanitized_column(sort)
